chore: improve Collapse  exports
